+++ tests/alerters/teams_test.py
@@ -36,7 +36,7 @@ def test_ms_teams(caplog):
         '@context': 'http://schema.org/extensions',
         'summary': rule['ms_teams_alert_summary'],
         'title': rule['alert_subject'],
-        'text': BasicMatchString(rule, match).__str__()
+        'sections': [{'text': BasicMatchString(rule, match).__str__()}]
     }
     mock_post_request.assert_called_once_with(
         rule['ms_teams_webhook_url'],
@@ -78,7 +78,7 @@ def test_ms_teams_uses_color_and_fixed_width_text():
         'summary': rule['ms_teams_alert_summary'],
         'title': rule['alert_subject'],
         'themeColor': '#124578',
-        'text': body
+        'sections': [{'text': body}]
     }
     mock_post_request.assert_called_once_with(
         rule['ms_teams_webhook_url'],
@@ -115,7 +115,7 @@ def test_ms_teams_proxy():
         '@context': 'http://schema.org/extensions',
         'summary': rule['ms_teams_alert_summary'],
         'title': rule['alert_subject'],
-        'text': BasicMatchString(rule, match).__str__()
+        'sections': [{'text': BasicMatchString(rule, match).__str__()}]
     }
     mock_post_request.assert_called_once_with(
         rule['ms_teams_webhook_url'],
@@ -147,7 +147,7 @@ def test_ms_teams_ea_exception():
         mock_run = mock.MagicMock(side_effect=RequestException)
         with mock.patch('requests.post', mock_run), pytest.raises(RequestException):
             alert.alert([match])
-    assert 'Error posting to ms teams: ' in str(ea)
+    assert 'Error posting to MS Teams: ' in str(ea)
 
 
 def test_ms_teams_getinfo():
@@ -240,7 +240,7 @@ def test_ms_teams_ca_certs(ca_certs, ignore_ssl_errors, excpet_verify):
         '@context': 'http://schema.org/extensions',
         'summary': rule['ms_teams_alert_summary'],
         'title': rule['alert_subject'],
-        'text': BasicMatchString(rule, match).__str__()
+        'sections': [{'text': BasicMatchString(rule, match).__str__()}]
     }
     mock_post_request.assert_called_once_with(
         rule['ms_teams_webhook_url'],
@@ -248,5 +248,198 @@ def test_ms_teams_ca_certs(ca_certs, ignore_ssl_errors, excpet_verify):
         headers={'content-type': 'application/json'},
         proxies=None,
         verify=excpet_verify
+    )
+    assert expected_data == json.loads(mock_post_request.call_args_list[0][1]['data'])
+
+
+def test_ms_teams_attach_kibana_discover_url_when_generated():
+    rule = {
+        'name': 'Test Rule',
+        'type': 'any',
+        'ms_teams_attach_kibana_discover_url': True,
+        'ms_teams_webhook_url': 'http://test.webhook.url',
+        'ms_teams_alert_summary': 'Alert from ElastAlert',
+        'alert': [],
+        'alert_subject': 'Cool subject',
+    }
+    rules_loader = FileRulesLoader({})
+    rules_loader.load_modules(rule)
+    alert = MsTeamsAlerter(rule)
+    match = {
+        '@timestamp': '2016-01-01T00:00:00',
+        'kibana_discover_url': 'http://kibana#discover'
+    }
+    with mock.patch('requests.post') as mock_post_request:
+        alert.alert([match])
+
+    expected_data = {
+        '@type': 'MessageCard',
+        '@context': 'http://schema.org/extensions',
+        'summary': rule['ms_teams_alert_summary'],
+        'title': rule['alert_subject'],
+        'sections': [{'text': BasicMatchString(rule, match).__str__()}],
+        'potentialAction': [
+            {
+                '@type': 'OpenUri',
+                'name': 'Discover in Kibana',
+                'targets': [
+                    {
+                        'os': 'default',
+                        'uri': 'http://kibana#discover',
+                    }
+                ],
+            }
+        ],
+    }
+    mock_post_request.assert_called_once_with(
+        rule['ms_teams_webhook_url'],
+        data=mock.ANY,
+        headers={'content-type': 'application/json'},
+        proxies=None,
+        verify=True
+    )
+    actual_data = json.loads(mock_post_request.call_args_list[0][1]['data'])
+    assert expected_data == actual_data
+
+
+def test_ms_teams_attach_kibana_discover_url_when_not_generated():
+    rule = {
+        'name': 'Test Rule',
+        'type': 'any',
+        'ms_teams_attach_kibana_discover_url': True,
+        'ms_teams_webhook_url': 'http://test.webhook.url',
+        'ms_teams_alert_summary': 'Alert from ElastAlert',
+        'alert': [],
+        'alert_subject': 'Cool subject',
+    }
+    rules_loader = FileRulesLoader({})
+    rules_loader.load_modules(rule)
+    alert = MsTeamsAlerter(rule)
+    match = {
+        '@timestamp': '2016-01-01T00:00:00'
+    }
+    with mock.patch('requests.post') as mock_post_request:
+        alert.alert([match])
+
+    expected_data = {
+        '@type': 'MessageCard',
+        '@context': 'http://schema.org/extensions',
+        'summary': rule['ms_teams_alert_summary'],
+        'title': rule['alert_subject'],
+        'sections': [{'text': BasicMatchString(rule, match).__str__()}],
+    }
+    mock_post_request.assert_called_once_with(
+        rule['ms_teams_webhook_url'],
+        data=mock.ANY,
+        headers={'content-type': 'application/json'},
+        proxies=None,
+        verify=True
+    )
+    actual_data = json.loads(mock_post_request.call_args_list[0][1]['data'])
+    assert expected_data == actual_data
+
+
+def test_ms_teams_kibana_discover_title():
+    rule = {
+        'name': 'Test Rule',
+        'type': 'any',
+        'ms_teams_attach_kibana_discover_url': True,
+        'ms_teams_kibana_discover_title': 'Click to discover in Kibana',
+        'ms_teams_webhook_url': 'http://test.webhook.url',
+        'ms_teams_alert_summary': 'Alert from ElastAlert',
+        'alert': [],
+        'alert_subject': 'Cool subject',
+    }
+    rules_loader = FileRulesLoader({})
+    rules_loader.load_modules(rule)
+    alert = MsTeamsAlerter(rule)
+    match = {
+        '@timestamp': '2016-01-01T00:00:00',
+        'kibana_discover_url': 'http://kibana#discover'
+    }
+    with mock.patch('requests.post') as mock_post_request:
+        alert.alert([match])
+
+    expected_data = {
+        '@type': 'MessageCard',
+        '@context': 'http://schema.org/extensions',
+        'summary': rule['ms_teams_alert_summary'],
+        'title': rule['alert_subject'],
+        'sections': [{'text': BasicMatchString(rule, match).__str__()}],
+        'potentialAction': [
+            {
+                '@type': 'OpenUri',
+                'name': 'Click to discover in Kibana',
+                'targets': [
+                    {
+                        'os': 'default',
+                        'uri': 'http://kibana#discover',
+                    }
+                ],
+            }
+        ],
+    }
+    mock_post_request.assert_called_once_with(
+        rule['ms_teams_webhook_url'],
+        data=mock.ANY,
+        headers={'content-type': 'application/json'},
+        proxies=None,
+        verify=True
+    )
+    actual_data = json.loads(mock_post_request.call_args_list[0][1]['data'])
+    assert expected_data == actual_data
+
+
+def test_ms_teams_alert_facts():
+    rule = {
+        'name': 'Test Rule',
+        'type': 'any',
+        'ms_teams_webhook_url': 'http://test.webhook.url',
+        'ms_teams_alert_summary': 'Alert from ElastAlert',
+        'ms_teams_alert_facts': [
+            {
+                'name': 'Host',
+                'value': 'somefield',
+            },
+            {
+                'name': 'Sensors',
+                'value': '@timestamp',
+            }
+        ],
+        'alert_subject': 'Cool subject',
+        'alert': []
+    }
+    rules_loader = FileRulesLoader({})
+    rules_loader.load_modules(rule)
+    alert = MsTeamsAlerter(rule)
+    match = {
+        '@timestamp': '2016-01-01T00:00:00',
+        'somefield': 'foobarbaz'
+    }
+    with mock.patch('requests.post') as mock_post_request:
+        alert.alert([match])
+
+    expected_data = {
+        '@type': 'MessageCard',
+        '@context': 'http://schema.org/extensions',
+        'summary': rule['ms_teams_alert_summary'],
+        'title': rule['alert_subject'],
+        'sections': [
+            {
+                'text': BasicMatchString(rule, match).__str__(),
+                'facts': [
+                    {'name': 'Host', 'value': 'foobarbaz'},
+                    {'name': 'Sensors', 'value': '2016-01-01T00:00:00'},
+                ],
+            }
+        ],
+    }
+
+    mock_post_request.assert_called_once_with(
+        rule['ms_teams_webhook_url'],
+        data=mock.ANY,
+        headers={'content-type': 'application/json'},
+        proxies=None,
+        verify=True
     )
     assert expected_data == json.loads(mock_post_request.call_args_list[0][1]['data'])